+++ src/TransactionProcessor.java
@@ -1,11 +1,10 @@
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Scanner;
 
 public class TransactionProcessor {
-    public void processTransaction(List<BankAccount> accounts, double ammount){
-//        Scanner in = new Scanner(System.in);
-        for(BankAccount account: accounts){
-//            ammount=in.nextInt();
+    public void processTransaction(List<BankAccount> accounts, BigDecimal ammount) {
+        for (BankAccount account : accounts) {
             account.withdraw(ammount);
             System.out.println(account.getBalance());
         }
